Initial commit
This contains code that uses light intensity of the LED button on
top of the Google AIY Voice Kit v1 to display audio intensity being
picked up by the microphone. It is not compatible with version 2
of the Google AIY Voice Kit, which uses an RGB LED in the button.